+++ Engine/Shared/Interfaces/IMapApi.cs
@@ -5,8 +5,8 @@ namespace Shared;
     public interface IMapApi
     {
         IMap? GetMap();
-        ITurf? GetTurf(int x, int y, int z);
-        void SetTurf(int x, int y, int z, int turfId);
+        ITurf? GetTurf(long x, long y, long z);
+        void SetTurf(long x, long y, long z, int turfId);
         Task<IMap?> LoadMapAsync(string filePath);
         void SetMap(IMap map);
         Task SaveMapAsync(string filePath);
@@ -14,20 +14,20 @@ namespace Shared;
         /// <summary>
         /// Gets all game objects within a specified range of a central point.
         /// </summary>
-        IEnumerable<IGameObject> GetObjectsInRange(int x, int y, int z, int range);
+        IEnumerable<IGameObject> GetObjectsInRange(long x, long y, long z, int range);
 
         /// <summary>
         /// Gets all game objects of a specific type (and its subtypes) within a specified range.
         /// </summary>
-        IEnumerable<IGameObject> GetObjectsInRange(int x, int y, int z, int range, string typePath);
+        IEnumerable<IGameObject> GetObjectsInRange(long x, long y, long z, int range, string typePath);
 
         /// <summary>
         /// Gets all game objects within a specified rectangular area.
         /// </summary>
-        IEnumerable<IGameObject> GetObjectsInArea(int x1, int y1, int x2, int y2, int z);
+        IEnumerable<IGameObject> GetObjectsInArea(long x1, long y1, long x2, long y2, long z);
 
         /// <summary>
         /// Gets all game objects of a specific type (and its subtypes) within a specified rectangular area.
         /// </summary>
-        IEnumerable<IGameObject> GetObjectsInArea(int x1, int y1, int x2, int y2, int z, string typePath);
+        IEnumerable<IGameObject> GetObjectsInArea(long x1, long y1, long x2, long y2, long z, string typePath);
     }
