+++ Engine/Shared/Interfaces/IRegionApi.cs
@@ -1,6 +1,6 @@
 namespace Shared;
     public interface IRegionApi
     {
-        void SetRegionActive(int x, int y, int z, bool active);
-        bool IsRegionActive(int x, int y, int z);
+        void SetRegionActive(long x, long y, long z, bool active);
+        bool IsRegionActive(long x, long y, long z);
     }
