+++ Engine/Shared/Interfaces/IStandardLibraryApi.cs
@@ -5,7 +5,7 @@ namespace Shared;
     {
         GameObject? Locate(string typePath, List<GameObject> container);
         void Sleep(int milliseconds);
-        List<GameObject> Range(int distance, int centerX, int centerY, int centerZ);
+        List<GameObject> Range(int distance, long centerX, long centerY, long centerZ);
         List<GameObject> View(int distance, GameObject viewer);
         int Step(GameObject obj, int dir, int speed);
         int StepTo(GameObject obj, GameObject target, int minDist, int speed);
